// src/services/content/interface/cs-content-type.ts
export enum CsContentType {
    ASSET = 'Asset',
    CLASSROOM_TEACHING_VIDEO = 'ClassroomTeachingVideo',
    CONCEPT_MAP = 'ConceptMap',
    COURSE = 'Course',
    COURSE_UNIT = 'CourseUnit',
    CURIOSITY_QUESTION_SET = 'CuriosityQuestionSet',
    E_TEXTBOOK = 'eTextBook',
    EXPERIENTIAL_RESOURCE = 'ExperientialResource',
    EXPLANATION_RESOURCE = 'ExplanationResource',
    EXPLANATION_VIDEO = 'Explanation Video',
    FOCUS_SPOT = 'FocusSpot',
    LEARNING_OUTCOME_DEFINITION = 'LearningOutcomeDefinition',
    LESSON_PLAN_UNIT = 'LessonPlanUnit',
    MARKING_SCHEME_RUBRIC = 'MarkingSchemeRubric',
    ONBOARDING_RESOURCE = 'OnboardingResource',
    PEDAGOGY_FLOW = 'PedagogyFlow',
    PLUGIN = 'Plugin',
    PRACTICE_QUESTION_SET = 'PracticeQuestionSet',
    PRACTICE_RESOURCE = 'PracticeResource',
    READING_MATERIAL = 'ReadingMaterial',
    SELF_ASSESS = 'SelfAssess',
    TEACHING_METHOD = 'TeachingMethod',
    TEMPLATE = 'Template',
    TEXTBOOK = 'TextBook',
    TEXTBOOK_UNIT = 'TextBookUnit',
    COLLECTION = 'Collection',
    EXPLANATION_READING_MATERIAL = 'ExplanationReadingMaterial',
    LEARNING_ACTIVITY = 'LearningActivity',
    LESSON_PLAN = 'LessonPlan',
    LESSON_PLAN_RESOURCE = 'LessonPlanResource',
    PREVIOUS_BOARD_EXAM_PAPERS = 'PreviousBoardExamPapers',
    TV_LESSION = 'TVLesson',
    RESOURCE = 'Resource'
}